Refactoring the USER REDUX stuff
