+++ frontend/src/Redux/reducers.js
@@ -3,10 +3,10 @@ import houseIssues from '../Dashboard/reducers';
 import house from '../House/reducers';
 import issue from '../Issue/reducers';
 import search from '../Search/reducers';
-import user from './userReducer';
+import { userReducer } from '../User';
 
 export default combineReducers({
-  user,
+  userReducer,
   houseIssues,
   house,
   issue,
