+++ frontend/src/Routes/Routes.js
@@ -53,7 +53,7 @@ class Routes extends Component {
 
 const mapStateToProps = state => {
   return {
-    user: state.user.user,
+    user: state.userReducer.user,
   };
 };
