+++ frontend/src/User/index.js
@@ -2,5 +2,6 @@ import Login from './Login';
 import Register from './Register';
 import Forgot from './Forgot';
 import Logout from './Logout';
+import userReducer from './reducer';
 
-export { Login, Logout, Register, Forgot };
+export { Login, Logout, Register, Forgot , userReducer};
